Just take a ConnectInfo here.
We use `newWorkers` in the creation of our foundation data type, and
therefore we can't pass in the foundation record - a catch 22.

Instead, we just ask for a ConnectInfo. As this will typically come from
the app settings, and instantiation of the foundation record is done
with the app settings, we have this easily available to pass in.